Change some string property assessment columns to booleans (#140)
* Add migration to make own_occ boolean

This migration changes the definition of the property_assessments own_occ
column to be boolean. It DOES NOT migrate existing imported data -- so if we
had already uploaded data with "Y" or "N" values, this may not be converted
correctly.

* Add migration to make corner_unit boolean

This migration changes the definition of the property_assessments corner_unit
column to be boolean. It DOES NOT migrate existing imported data -- so if we
had already uploaded data with "Y" or "N" values, this may not be converted
correctly.

* Run rubocop