Support creating new homes when filling out survey (#341)
* Add migration for homes.canonicalized attribute
* Add todos for canonicalizing home model
* Set canonicalized flag when creating/updating homes
* Add dependency on sucker_punch
* Create canonicalize address job and wire up

Create an Active Job to canonicalize home addresses, after these are
created or changed. This currently has a dummy implementation -- would
be good to factor out this logic into a separate module.

* Canonicalize addresses when first imported

* Fix rubocop violations

* Add comment

* Move canonicalization logic to model using callbacks

* Do not manually canonicalize addresses when seeded

* Make surveyor optional on SurveyVisit

To support anonymous surveys, we want to allow survey visits to have a null
surveyor, so cannot be belongs_to.

* Update canonicalize job to reassign survey visits

* Add migration to store whether home is added by anon user

* Change home canonicalized attribute to status

Rather than having two attributes for this, we have a single string. New
homes start with status = 'uncanonicalized'. When this is validated and
updated, it is updated to 'canonicalized. If the address lookup fails,
it is updated to 'unrecognized'.

* Switch to using enum for status and fix logic

* This updates Mike’s class such that it will now actually Canonicalize the address and call Pelias. 🎉

Make home.visited? and .completed? methods not be protected.

---------

Co-authored-by: Matthew Zagaja <mzagaja@alum.wpi.edu>
Co-authored-by: Linden Huhmann <lhuhmann@users.noreply.github.com>
Co-authored-by: Anna Westland-Tegart <aTegart@users.noreply.github.com>
Co-authored-by: Matthew Zagaja <mzagaja@users.noreply.github.com>