Resolve #657 Migrate to Survey Answers (#669)
* Resolve #657 Migrate to Survey Answers

Now that we're using an array of survey answers instead of a singular answer, we need to update our exporter to comport with this change, and update our tests like proper software engineers.

* Bump Ruby

Appease the rubocop.

* Use Safer Ruby

* Ruby 3.4.x seems to have some potential bugs with Rails 7.1 per Laura's link at https://www.fastruby.io/blog/ruby/rails/versions/compatibility-table.html. So we retreat to latest in 3.2 series for now.

* Remove syntax_suggest

This is now a default gem and was making CI angry.

* Bump Rack

We're insecure. don't know what for.